list底部布局bugfix

+++ frameworks/core/components_ng/pattern/list/list_item_group_layout_algorithm.h
@@ -388,7 +388,7 @@ public:
         prevTotalItemCount_ = prevTotalItemCount;
     }
 
-    void SetPrevTotalMainSize(int32_t prevTotalMainSize)
+    void SetPrevTotalMainSize(float prevTotalMainSize)
     {
         prevTotalMainSize_ = prevTotalMainSize;
     }
